requirements.txt
dspy-ai>=2.4.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
openai>=1.0.0
